+++ lib/flex_widget.dart
@@ -1,32 +1,62 @@
+// Copyright 2019 Marvin Quevedo. All rights reserved.
+// Use of this source code is governed by a Apache 2.0 license that can be
+// found in the LICENSE file.
+
 import 'package:flutter/widgets.dart';
 
 import 'responsive.dart';
 
+// ignore: must_be_immutable
 class FlexWidget extends StatelessWidget {
+  /// Child for apply the flex Columns
   final Widget child;
-  //Columns counts for less than 768
+
+  /// Columns counts for less than 768
   final double xs;
+
+  /// Columns offset for less than 768
   final double xsOffset;
+
+  /// Columns counts for less than 768 and the device is landscape
   double xsLand;
+
+  /// Columns offset for less than 768 and the device is landscape
   double xsLandOffset;
 
-  //Columns counts for more or equals to 768
+  /// Columns counts for more or equals to 768 pixels
   final double sm;
+
+  /// Columns offset for more or equals to 768 pixels
   final double smOffset;
 
+  /// Columns counts for more or equals to 768 pixels and the device is landscape
   double smLand;
-  double smLandOffset;
-  //Columns counts for more or equals to 992
-  final double md;
-  final double mdOffsset;
 
+  /// Columns offset for more or equals to 768 pixels and the device is landscape
+  double smLandOffset;
+
+  ///Columns counts for more or equals to 992 pixels
+  final double md;
+
+  /// Columns offset for more or equals to 992 pixels
+  final double mdOffset;
+
+  /// Columns counts for more or equals to 992 pixels and the device is landscape
   double mdLand;
+
+  /// Columns offset for more or equals to 992 pixels and the device is landscape
   double mdLandOffset;
-  //Columns counts for more or equals to 1170
+
+  ///Columns counts for more or equals to 1170 pixels
   final double lg;
+
+  /// Columns offset for more or equals to 1170 pixels
   final double lgOffset;
 
+  /// Columns counts for more or equals to 1170 pixels and the device is landscape
   double lgLand;
+
+  /// Columns offset for more or equals to 1170 pixels and the device is landscape
   double lgLandOffset;
 
   FlexWidget(
@@ -40,15 +70,13 @@ class FlexWidget extends StatelessWidget {
       this.smLand,
       this.smLandOffset,
       this.md = 6,
-      this.mdOffsset = 0,
+      this.mdOffset = 0,
       this.mdLand,
       this.mdLandOffset,
       this.lg = 3,
       this.lgOffset = 0,
       this.lgLand,
       this.lgLandOffset}) {
-
-    // C
     if (xsLand == null) {
       xsLand = xs;
     }
@@ -70,7 +98,7 @@ class FlexWidget extends StatelessWidget {
     }
 
     if (mdLandOffset == null) {
-      mdLandOffset = mdOffsset;
+      mdLandOffset = mdOffset;
     }
     if (lgLandOffset == null) {
       lgLandOffset = lgOffset;
@@ -82,21 +110,20 @@ class FlexWidget extends StatelessWidget {
     //obtain the screen size
     final size = MediaQuery.of(context).size;
     final orientation = MediaQuery.of(context).orientation;
-
-    final columnsCount = ResponsiveSettings.of(context).columnsCount;
+    final settings = ResponsiveSettings.of(context);
+    if (settings == null) {
+      throw Exception("The FlexWidget is not into a ResponsiveRow");
+    }
+    final columnsCount = settings.columnsCount;
 
     if (lg > columnsCount) {
-      throw Exception(
-          "lg is bigger than columnsCount = ${lg}  > ${columnsCount}");
+      throw Exception("lg is bigger than columnsCount = $lg  > $columnsCount");
     } else if (md > columnsCount) {
-      throw Exception(
-          "md is bigger than columnsCount = ${md}  > ${columnsCount}");
+      throw Exception("md is bigger than columnsCount = $md  > $columnsCount");
     } else if (sm > columnsCount) {
-      throw Exception(
-          "sm is bigger than columnsCount = ${sm}  > ${columnsCount}");
+      throw Exception("sm is bigger than columnsCount = $sm  > $columnsCount");
     } else if (xs > columnsCount) {
-      throw Exception(
-          "xs is bigger than columnsCount = ${xs}  > ${columnsCount}");
+      throw Exception("xs is bigger than columnsCount = $xs  > $columnsCount");
     }
 
     //use the OrientationBuilder Widget
@@ -122,12 +149,11 @@ class FlexWidget extends StatelessWidget {
     }
   }
 
+  /// Calc the width in pixels, for the screen and rotation config used actually in the device
   double _calcWidth(Size size, Orientation orientation, columnsCount) {
-
     final width = size.width;
     if (orientation == Orientation.portrait) {
-
-      print("width = ${width}");
+      //print("width = $width");
       final colWidth = width / columnsCount;
       if (colWidth > 1170) {
         // print("lg");
@@ -145,7 +171,7 @@ class FlexWidget extends StatelessWidget {
     } else {
       //Landscape configuration
 
-      print("Height = ${width}");
+      // print("Height = $width");
       final colWidth = width / columnsCount;
       if (colWidth > 1170) {
         // print("lg");
@@ -165,26 +191,20 @@ class FlexWidget extends StatelessWidget {
 
   double _calcOffset(Size size, Orientation orientation, columnsCount) {
     if (orientation == Orientation.portrait) {
-      print("portrait");
       final width = size.width;
 
       final colWidth = width / columnsCount;
       if (colWidth > 1170) {
-        // print("lg");
         return colWidth * lgOffset;
       } else if (width >= 992) {
-        //print("md");
-        return colWidth * mdOffsset;
+        return colWidth * mdOffset;
       } else if (width >= 768) {
-        //print("sm");
         return colWidth * smOffset;
       } else {
-        //print("xs");
         return colWidth * xsOffset;
       }
     } else {
-      print("landscape");
-      //Lanscape configuration
+      //Landscape configuration
       final width = size.height;
       final colWidth = width / columnsCount;
       if (colWidth > 1170) {